Add Controller

+++ app/Http/Controllers/Backsite/DoctorController.php
@@ -4,19 +4,23 @@ namespace App\Http\Controllers\Backsite;
 
 use App\Http\Controllers\Controller;
 
-//use library
+// use library here
 use Illuminate\Support\Facades\Storage;
+use Illuminate\Support\Str;
+use Illuminate\Http\Request;
 use Symfony\Component\HttpFoundation\Response;
 
 //Request
-use App\Requests\Doctor\Backsite\StoreDoctorRequest;
-use App\Requests\Doctor\Backsite\UpdateDoctorRequest;
+use App\Http\Requests\Doctor\StoreDoctorRequest;
+use App\Http\Requests\Doctor\UpdateDoctorRequest;
 
-//use everything
-//use Gate;
+// use everything here
+use Gate;
 use Auth;
+use File;
 
-//model here
+// use model here
+use App\Models\User;
 use App\Models\Operational\Doctor;
 use App\Models\MasterData\Specialist;
 
@@ -35,13 +39,18 @@ class DoctorController extends Controller
      */
     public function index()
     {
-        //for table grid
+        // abort_if(Gate::denies('doctor_access'), Response::HTTP_FORBIDDEN, '403 Forbidden');
+
+        // for table grid
         $doctor = Doctor::orderBy('created_at', 'desc')->get();
 
-        //for select2 = ascending a to z
+        // for select2 = ascending a to z
         $specialist = Specialist::orderBy('name', 'asc')->get();
+        $user = User::whereHas('detail_user', function($query){
+                        $query->where('type_user_id', 2);
+                    })->orderBy('name', 'asc')->get();
 
-        return view('pages.backsite.operational.doctor.index', compact('doctor','specialist'));
+        return view('pages.backsite.operational.doctor.index', compact('doctor', 'specialist', 'user'));
     }
 
     /**
@@ -57,10 +66,29 @@ class DoctorController extends Controller
      */
     public function store(StoreDoctorRequest $request)
     {
-        //Get all Request from frontsite
+        // get all request from frontsite
         $data = $request->all();
 
-        //Store to database
+        // re format before push to table
+        $data['fee'] = str_replace(',', '', $data['fee']);
+        $data['fee'] = str_replace('IDR ', '', $data['fee']);
+
+        // upload process here
+        $path = public_path('app/public/assets/file-doctor');
+        if(!File::isDirectory($path)){
+            $response = Storage::makeDirectory('public/assets/file-doctor');
+        }
+
+        // change file locations
+        if(isset($data['photo'])){
+            $data['photo'] = $request->file('photo')->store(
+                'assets/file-doctor', 'public'
+            );
+        }else{
+            $data['photo'] = "";
+        }
+
+        // store to database
         $doctor = Doctor::create($data);
 
         alert()->success('Success Message', 'Successfully added new doctor');
@@ -72,6 +100,8 @@ class DoctorController extends Controller
      */
     public function show(Doctor $doctor)
     {
+        abort_if(Gate::denies('doctor_show'), Response::HTTP_FORBIDDEN, '403 Forbidden');
+
         return view('pages.backsite.operational.doctor.show', compact('doctor'));
     }
 
@@ -81,10 +111,15 @@ class DoctorController extends Controller
      */
     public function edit(Doctor $doctor)
     {
-        //for select2 = ascending a to z
-        $specialist = Specialist::orderBy('name', 'asc')->get();
+        abort_if(Gate::denies('doctor_edit'), Response::HTTP_FORBIDDEN, '403 Forbidden');
 
-        return view('pages.backsite.operational.doctor.edit', compact('doctor', 'specialist'));
+        // for select2 = ascending a to z
+        $specialist = Specialist::orderBy('name', 'asc')->get();
+        $user = User::whereHas('detail_user', function($query){
+                        $query->where('type_user_id', 2);
+                    })->orderBy('name', 'asc')->get();
+
+        return view('pages.backsite.operational.doctor.edit', compact('doctor', 'specialist', 'user'));
     }
 
     /**
@@ -92,10 +127,36 @@ class DoctorController extends Controller
      */
     public function update(UpdateDoctorRequest $request, Doctor $doctor)
     {
-        //get all request from frontsite
+        // get all request from frontsite
         $data = $request->all();
 
-        //update to database
+        // re format before push to table
+        $data['fee'] = str_replace(',', '', $data['fee']);
+        $data['fee'] = str_replace('IDR ', '', $data['fee']);
+
+        // upload process here
+        // change format photo
+        if(isset($data['photo'])){
+
+             // first checking old photo to delete from storage
+            $get_item = $doctor['photo'];
+
+            // change file locations
+            $data['photo'] = $request->file('photo')->store(
+                'assets/file-doctor', 'public'
+            );
+
+            // delete old photo from storage
+            $data_old = 'storage/'.$get_item;
+            if (File::exists($data_old)) {
+                File::delete($data_old);
+            }else{
+                File::delete('storage/app/public/'.$get_item);
+            }
+
+        }
+
+        // update to database
         $doctor->update($data);
 
         alert()->success('Success Message', 'Successfully updated doctor');
@@ -107,9 +168,21 @@ class DoctorController extends Controller
      */
     public function destroy(Doctor $doctor)
     {
+        abort_if(Gate::denies('doctor_delete'), Response::HTTP_FORBIDDEN, '403 Forbidden');
+
+        // first checking old file to delete from storage
+        $get_item = $doctor['photo'];
+
+        $data = 'storage/'.$get_item;
+        if (File::exists($data)) {
+            File::delete($data);
+        }else{
+            File::delete('storage/app/public/'.$get_item);
+        }
+
         $doctor->forceDelete();
 
         alert()->success('Success Message', 'Successfully deleted doctor');
-        return back();   
+        return back();  
     }
 }
